+++ mne_bids/dig.py
@@ -304,7 +304,8 @@ def _write_coordsystem_json(*, raw, unit, hpi_coord_system,
     _write_json(fname, fid_json, overwrite=True, verbose=verbose)
 
 
-def _write_dig_bids(bids_path, raw, overwrite=False, verbose=True):
+def _write_dig_bids(bids_path, raw, montage=None, acpc_aligned=False,
+                    overwrite=False, verbose=True):
     """Write BIDS formatted DigMontage from Raw instance.
 
     Handles coordinatesystem.json and electrodes.tsv writing
@@ -319,6 +320,11 @@ def _write_dig_bids(bids_path, raw, overwrite=False, verbose=True):
         data, ``electrodes.tsv`` are not saved.
     raw : mne.io.Raw
         The data as MNE-Python Raw object.
+    montage : mne.channels.DigMontage | None
+        The montage to use rather than the one in ``raw`` if it
+        must be transformed from the "head" coordinate frame.
+    acpc_aligned : bool
+        Whether "mri" space is aligned to ACPC.
     overwrite : bool
         Whether to overwrite the existing file.
         Defaults to False.
@@ -328,10 +334,19 @@ def _write_dig_bids(bids_path, raw, overwrite=False, verbose=True):
     # write electrodes data for iEEG and EEG
     unit = "m"  # defaults to meters
 
+    if montage is None:
+        montage = raw.get_montage()
+    else:
+        # prevent transformation back to "head", only should be used
+        # in this specific circumstance
+        if bids_path.datatype == 'ieeg':
+            montage.remove_fiducials()
+        raw.set_montage(montage)
+
     # get coordinate frame from digMontage
-    digpoint = raw.info['dig'][0]
+    digpoint = montage.dig[0]
     if any(digpoint['coord_frame'] != _digpoint['coord_frame']
-           for _digpoint in raw.info['dig']):
+           for _digpoint in montage.dig):
         warn("Not all digpoints have the same coordinate frame. "
              "Skipping electrodes.tsv writing...")
         return
@@ -340,6 +355,17 @@ def _write_dig_bids(bids_path, raw, overwrite=False, verbose=True):
     coord_frame_int = int(digpoint['coord_frame'])
     mne_coord_frame = MNE_FRAME_TO_STR.get(coord_frame_int, None)
     coord_frame = MNE_TO_BIDS_FRAMES.get(mne_coord_frame, None)
+
+    if bids_path.datatype == 'ieeg' and mne_coord_frame == 'mri':
+        if acpc_aligned:
+            coord_frame = 'ACPC'
+        else:
+            raise RuntimeError(
+                '`acpc_aligned` is False, if your T1 is not aligned '
+                'to ACPC and the coordinates are in fact in ACPC '
+                'space there will be no way to relate the coordinates '
+                'to the T1. If the T1 is ACPC-aligned, use '
+                '`acpc_aligned=True`')
 
     # create electrodes/coordsystem files using a subset of entities
     # that are specified for these files in the specification